Reset ATTR_SPCWKUP_COUNT to 0 at start of MPIPL
We were seeing issues in the newly added PM_RESET call in istep 21.
This was happening because if the SKIP_WAKEUP attribute is not set
then PM_RESET will attempt to enable special_wakeup on all of the
EX targets before enabling auto-special wakeup. We are then expecting
special wakeup done bit to be set on all of th EX targets prior
to enabling auto-special wakeup on the target. IF SPCWKUP_COUNT is
still set to a non-zero value from the previous IPL the call to
enable special_wakeup will result in a no-op so the special wkup
done bit will never be set in the GPMMR. This causes to call to
enable auto-special wakeup  to fail

Change-Id: Iabdb5099ea9d6eb195003697de8ca074804d2d98
CQ: SW446135
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/66430
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>