Close windows where Hostboot HRMOR is not available to FSP
There are certain small time windows where the core scratch register, which
holds the Hostboot load address for FSP to read in order to dump Hostboot, is
cleared during winkle.  This change caches the Hostboot load address in a
processor attribute that FSP can read/use in the event it happens to request a
dump in the window where the core scratch register is not valid.

Change-Id: I776df5e64f4d8ad382392662d08ee349661ba78a
CQ: SW444212
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/66664
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>