ipmi: Terminate SEL task via shutdown event
The IpmiSEL task will become part of the extended image once the IPMI
module is split in two. Once split, if we need to handle an early shutdown
we cannot be referencing code from HBI in HBB. To that end, ensure we
don't instantiate an IpmiSEL via Singleton in an attempt to shutdown its
event loop when it hasn't been loaded, let alone started. Instead, have
IpmiSEL register itself in the shutdown handler, and shut down the SEL
task before IpmiRP.

Change-Id: I358f6cb1f5528a4ad72c93477ad883cad19e2bf6
Signed-off-by: Andrew Jeffery <andrewrj@au1.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/67076
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Corey V. Swenson <cswenson@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/67392