Skip resource recovery for node and power gard.
After the node is deconfigured with non-fatal or power
error running a resource recovery will always recover the node
since all vital parts are deconfigured in the node. so
skipping the resource recovery for node. Power gard need
be considered as fatal so skipping resource recovery for
power gard too.

Change-Id: I246733b59d908a56401a0c602897a5d12522406b
Signed-off-by: Dhruvaraj Subhashchandran <dhruvaraj@in.ibm.com>
CQ: SW444154
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/66981
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>