SMF: Port NVRAM Reading Logic From Skiboot
As part of SMF secure memory distribution, we need to be able
to read the size of secure memory from NVRAM PNOR partition
(for OpenPOWER). The functionality to read (index into) NVRAM
already exists in skiboot. This commits ports the bare minimum
functionality required to read NVRAM from skiboot to hb.

Change-Id: I17b9014ec3df590bcd8745ae70e0f96e36580117
RTC: 192411
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/67414
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Dean Sanner <dsanner@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Reviewed-by: Nicholas E. Bofferding <bofferdn@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>