Delay targeting update in CFM until all nodes are processed
Before rebooting HBRT onto a new level of code, we perform a
series of operations to run against the new level of targeting
data.  The new data is retrieved as a lid per node.

Previously we processed each node completely before moving to the
next one.  This is problematic because the act of applying the
new data to our live memory has the side effect of breaking some
future accesses to attributes that are required as part of the
lid processing.

The fix here defers all of the live memory updates until after
all of the lids have been processed, thus allowing attribute
lookups to continue to work.

Change-Id: I8e8aebb7f87afc5acbf8be86c0ec727f91fce64f
CQ: SW447768
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/67357
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>