Elevate log levels for simics during PSU ops
This commit introduces some new magic instructions one which will
temporarily elevate the log levels for given components and another
which will start and stop collection of these simics logs. This
was added so we can temporarily increase log levels during PSU
operations in hopes of catching a timeout we have been seeing
in simics and getting more info for the simics teams.

Change-Id: I990a4b5413f7ff14796dee36e39199f785aef458
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/67359
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Hieu C. Nguyen <hieu.nguyen@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>