Wrap TS_ macros in {} to avoid strange IF statement behavior
The TS_ macros will replace the macro with multiple lines of code.
So if you put a TS_FAIL in an if statement that was not enclosed
in {} you could get strange/unexpected behavior. This addes an extra
layer of protection to ensure that the macro is enclosed in its own
execution block.

Change-Id: I226ccef81095cf8511b30e147e59d1da4904cd5c
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/67373
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Corey V. Swenson <cswenson@us.ibm.com>
Reviewed-by: Thi N. Tran <thi@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>