Revert "sio: Add test for availability - LPC error tweak"
This reverts commit 627379aeaa27e30d66ebb0aecf218708d465162c, which
incorrectly attempted to account for LPC errors during testing for SIO
absence. The broken patch interprets the OPBM status as LPCHC status and
expects an LPCHC SYNC Abnormal error, however as it's the OPBM status
that indicates the error we never hit the LPCHC error path to populate
the errl with the error RC expected by the SIO driver.

Change-Id: Ib993d4a2b9b4e5018d9273a1c82f8b5c21ec9a25
Signed-off-by: Andrew Jeffery <andrewrj@au1.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/67459
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>