SMF: Distribute Secure Memory Among Procs
Logic for distributing the requested amount of SMF memory between
the procs with memory on the system. The algorithm attempts to
allocate the memory in power-of-two increments of 256MB (such
is the HW limitation) under each proc until the requested
amount is satisfied or until we run out of procs with memory.

Change-Id: Ica3e1706bdb731762a3daf07c11d889fc7b6367f
RTC: 192411
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/68128
Reviewed-by: Nicholas E. Bofferding <bofferdn@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>