Add way for developers to disable hb sim logging via env variable
There was a request to add a way to disable a new feature that was
added to allow hostboot to temporarily elevate and capture simics
component logs. When hostboot puts the magic instructions to enable
log capture and elevate log levels it causes log spam and adds to
IPL time in simics. Simics devs can now set DISABLE_HB_SIMICS_LOGS
= 1 and the magic instruction will look for that env var and skip
collecting or elevating the logs.

Change-Id: Ib31bad3485833f642533dfd52af9cf941a2772d6
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/67882
Reviewed-by: CALEB SCHLOSSIN <calebs@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>