Route scom operations on OCMB chips to exp_i2c_scom interface
Before we get OMI targets trained we must use i2c to access scom
registers on the OCMB chip. This commit does all of the plumbing so
when HWP calls getScom() on a OCMB target, the hostboot platform
recognizes this as a special scom and routes it to a new i2scom DD.
This device driver will truncate the scom address to 32 bits and
run the exp_i2c_putscom/exp_i2c_getscom interfaces to perform
the operation. Eventually we need to also support MMIO scoms to
the OCMB chip, the MMIO scoms will be used after the OMI training
is complete.

Change-Id: I0018cc8d25f74d1253b72c3112d3e344a4248416
RTC: 196806
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/67976
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>