Disable automatic collection of SIM logs with MAGIC_SET_LOG_LEVEL
While we were trying to trace down some PSU simics bugs we added
a new feature where HB can elevate log levels of simics components
and copy out simics traces to a file which can be used for FFDC.
We found the bug which prompted this and we are seeing issues where
attempting to set the log level on procs in node2 is problematic.
This commit makes it so no logs are captured unless the env var
ENABLE_HB_SIMICS_LOGS is set in simics env.

Change-Id: Icc049b86daf89ff037edb1b5a1cf99c9ce551a2d
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/68140
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: CALEB SCHLOSSIN <calebs@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>