Clear INT_CQ related firs after reseting INTRRP logic in HB
With multi-node systems during the MPIPL steps up until the
re-init of the intrrp the nodes will be running completely
independently. Because of this we have seen errors where node C
is still sending hypervisor traffic to node A while node A has
already torn down its interrupt logic. These errors are irrelevant
as we reset the interrupt logic anyways.

CQ: SW448121
Change-Id: I910ab80da662b6ad4fb4fbc680dbf2957d2736da
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/68182
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>
Reviewed-by: Dean Sanner <dsanner@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Zane C. Shelley <zshelle@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>