Route scom operations on OCMB chips to fapi2 mmio scom interface
After we get OMI targets trained we can communicate with the OCMB
chips via MMIO. This includes reading and writing scom registers on
the OCMB chip. This commit hooks up all of the plumming so when a
HWP calls fapi2::getScom/putScom on an OCMB chip hostboot will be
able to determine what functions to call to perform the operation.

Change-Id: I3ae6f8b4ad3128f61d886b1fdfbeea82b0c6e76e
RTC: 196806
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/68353
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>