Fix CAPP target XML in Axone simics xml file
There was some incorrect/missing information on the CAPP target
in the newly hand created simics Axone xml. The FAPI_POS attribute
for capp1 was set to 0, when it should be 1. Also the chiplet ID
was incorrect for capp1. The chiplet ID for capp1 in all p9 systems
should be 4.

Change-Id: I42392b2f7afd1827cdad15e6308750a055f8a859
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/68504
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>