Define Parent/Child Relationship for OMIC/OMI targets
This commit puts in the plumbing to have xmltohb.pl  look for
ATTR_OMIC_PARENT attributes on the OMI targets and with the value
of that attribute define a bi-directional relationship between the
given OMI target and its defined OMIC parent. Each target in the binary
will have pointers to its associated parent/child. When getChildren<OMI>
or getParent<OMIC> is called in the FAPI2 api for a OMIC or OMI target
respectively, then the Hostboot platform implementation of these
functions will route to the new getParentOmicTargetsByState and
getChildOmiTargetsByState functions that were defined to perform lookups
in the targeting binary for this relationship.

Change-Id: I8cd901864a700c9fe575dfa0916d5e78760a7b0c
RTC: 172969
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/68541
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Nicholas E. Bofferding <bofferdn@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>