Add Slave-Procs-Only SBE Update via FSI I2C in istep 6.9
This change adds the ability to call the SBE Update function and
have it run only against the slave processors in the system early
in the IPL when they can only access the SBE Seeproms via FSI I2C.
This ability is called in istep 6.9. There are some other minor
changes to support this, as well.

Change-Id: If48a66e6135ffdba8f60d80067f6015fbc3f10ff
CQ:SW423354
RTC:200280
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/69946
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Dean Sanner <dsanner@us.ibm.com>