Serialize the Check For Prologs in CI
Prolog checks are performed in parallel with build in hostboot
CI process, and the return code from the prolog checking
script is checked after the build job completes, which
introduces significant delays before the developer is informed
of incorrect prologs. This change serializes the check for
prolog correctness and forces the check to run before the build
process, so that the developer is informed about bad prologs
as soon as possible.

Change-Id: Ib84802f14022712268db1b0b539396e59d25792f
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/68678
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Christian R. Geddes <crgeddes@us.ibm.com>
Reviewed-by: Nicholas E. Bofferding <bofferdn@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>