Create error log and fail if Proc's EC is found to be 0x0
While looking up the SBE version we will look at the processor's
EC level to determine how to lookup the version. If we find that
the EC level is 0 then something went wrong. In HW the register
describing the procs EC will always be fused to some non-zero
so we shouldnt see this in HW , in simics if this happens it is
likely because simics has not implemented this register yet.

Change-Id: I24bc0caaf3d2c9a574943ca07069bd6fb99cf1a6
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/68804
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>