OCC active notification for NVDIMM protection
TMGT needs to let PHYP know when the NVDIMMs are protected
by the OCC.  This protection happens when the OCCs are in
active state.  When the OCC goes active, TMGT will send a
msg via HWSV->MBOX to HBRT.  HBRT will then notify PHYP of the
OCC protection status of the NVDIMMs.

Change-Id: Id23ab8020bd76941adcac76279cc46106dc2b1b2
RTC:201290
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/69577
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Christian R. Geddes <crgeddes@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>