Add expscom test cases
This commit introduces some tests that check that scoms out to the
explorer chip work. This first commit only enables test through the
i2c path to ddimm0. Eventually expscomtest.H should contain tests
that write and read over both i2c and mmio on ddimms that are and
are not behind mux chips.

Change-Id: I9a0e6612fd1d72a26991b936301aabd29508db90
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/69579
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>