Calculate capacity for 3DS DIMMs in IPMI FRU inventory
This commit fixes a bug in formula for calculating total capacity of
DIMM module that has memory ICs package with 3DS structure. In the
case of 3DS modules, it is necessary to take into account the die count
(byte 6 [bits 6: 4] in SPD) in the formula.
In addition, the duplicate entry in the Manufacturer list has
been removed.

Signed-off-by: Maxim Polyakov <m.polyakov@yadro.com>