Calculate capacity for 3DS DIMMs in IPMI FRU inventory
This commit fixes a bug in formula for calculating total capacity of
DIMM module that has memory ICs package with 3DS structure. In the
case of 3DS modules, it is necessary to take into account the die count
(byte 6 [bits 6: 4] in SPD) in the formula.
In addition, the duplicate entry in the Manufacturer list has
been removed.

Resolves #158
Signed-off-by: Maxim Polyakov <m.polyakov@yadro.com>
Change-Id: I3527684f14d833a203056464493f2dbfcdeec8ea
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/70055
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>