Fixed eeprom detection in the device tree
This commit fixed bug with determining the atmel compatible eeprom
in the device tree.
In addition, I added support for eeprom atmel-24c512 to the device
tree. Now the Hostboot sets the correct size (64Kb) of the mvpd,
cvpd, pvpd in the device tree for the i2c OS driver.

Signed-off-by: Maxim Polyakov <m.polyakov@yadro.com>