Fixed eeprom detection in the device tree
This commit fixed bug with determining the atmel compatible eeprom
in the device tree.
In addition, I added support for eeprom atmel-24c512 to the device
tree. Now the Hostboot sets the correct size (64Kb) of the mvpd,
cvpd, pvpd in the device tree for the i2c OS driver.

Resolves #160

Change-Id: I2dfccdda40f0dcf8159a353fdf8652c47542dd08
Signed-off-by: Maxim Polyakov <m.polyakov@yadro.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/70053
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>