Deconfigure parent if no more functioning children of child's type
Defect discovered that a DIMM deconfig did not rollup to the membuf
because the membuf also has an l4_centaur child.  The fix is
to check that the parent has no more children like the one being
deconfigured.  Also make sure that deconfiguring l4_centaur
does not deconfigure its parent membuf.

Change-Id: Ic499092d0da926dcd3e0c08f5205e491e04f9680
CQ:SW451955
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/70093
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Christian R. Geddes <crgeddes@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>