Add new pnorLayoutAxone.xml w/ new EECACHE section
This commit introduces a new pnor layout which will be used
when the simics_axone.config file is used. (Note: axone.config
was renamed to simics_axone.config). This new layout introduces
the EECACHE section which will be used to store copies of the
various EEPROMS in the system. The eventual goal is to be able to
remove the MVPD/DJVPD sections in PNOR and only use this EECACHE
section

Change-Id: Ifae610c4dd7f3aa9c87a5ca911cc4faa1ba2a98a
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/70172
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>