Fix EEPROM_VPD_PRIMARY_INFO attribute on proc targ in axone sim XML
Up until this point in P9 systems this attribute has had the
maxMemorySizeKB = 0x80 (128 KB) and the chipCount = 0x02. While
this is partially true, hostboot should never access the 2nd 64 KB
chip. The MVPD is completely stored in the first 64 KB chip. This
commit intentionally does not fix previous system XMLs but it might
be worth investigating. MRW is still supplying the old 0x80, 0x02 values
for Nimbus/Cumuls system so it was decided to leave old values in our
sim xmls for those systems.

RTC: 196805
Change-Id: Ibec7412359b6cda24a255ec612a5774a7ed3ac30
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/70259
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>