Remove sha512 hash from EECACHE section & init section w/ correct info
When the axone pnor layout was first introduced it mistakenly added
a sha-512 hash to the front of the EECACHE section. This will not work
because the section gets written to so the hash would not be very
helpful. Also because the section is ECC protected it cannot be
defaulted to all F's because that break the ECC rules. Instead the
pattern needs to be defaulted to have 8 bytes of 0xFF followed by
the ECC byte of 0x00. This follows what we did for previous ECC
protected sections.

Change-Id: I4dac3c097f5a7d959811ea1c1796231adec37cee
RTC: 196805
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/70472
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>