Add FSPBUILD to CI scripts
Added logic to the autoci scripts for a new FSPBUILD mode.  This
mode will perform the following actions:
- Build hostboot with the fsprelease.config file
- Build the errlparser code
- Not run simics

Also added support for a new env var (LOCAL_CI_USER) that can be
used to let you execute 'build-script' locally rather than
trying to use the 'hostboot' user id for some of the operations.

Change-Id: Ifc9d7875e00713cd46a27765aef5d24f88122a50
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/70617
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Camvan T. Nguyen <ctnguyen@us.ibm.com>
Reviewed-by: Sameer R. Veer <sveer@us.ibm.com>
Reviewed-by: Nicholas E. Bofferding <bofferdn@us.ibm.com>