Fix deconfigure parent rollup policy
Change to specifically designate if a target is
allowed to be deconfigured by child rollup and if
it should rollup to its parent.
Need to cover the case where there are two different types of children
and only one is allowed to rollup to the parent.
Also prevent the deconfigure rollup from happening to the parent.

Change-Id: I514876a46e9c8180e1fc99a969e0ca4247fbf2d9
CQ:SW454562
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/70759
Reviewed-by: Christian R. Geddes <crgeddes@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>